backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==2.0.0
redis==5.0.1
pinecone-client==2.2.4
google-generativeai==0.3.2
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
websockets==12.0
asyncio==3.4.3
langchain==0.1.0
langchain-google-genai==0.0.6
elevenlabs==0.2.26
celery==5.3.4
typing-extensions==4.8.0